Collapse "tool-search-maps".
